Add files via upload
MaedLedger Website for HTML / CSS /Js. This Website is a simple work for a company that is not real. But I made this project with the tutor to improve my HTML / CSS skillset. There is a mobile friendly css and wide_screen css file.  JavaScript used for Smooth Scrolling and map stuff.